SoundManager changes
Refactor SoundManager:
OneShot sounds are now cached and a OneShotProvider adds them to a mixer.
Music and looping sounds use a LoopingProvider (old name NAudioLibrary) to keep track of state.
